% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{default_registry}
\alias{default_registry}
\title{The registry `lmcc_bind`, `load_adapter` and `dump_adapter` use by default}
\usage{
default_registry()
}
\description{
The registry `lmcc_bind`, `load_adapter` and `dump_adapter` use by default
}
