% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{format_number}
\alias{format_number}
\title{Kernel section 7a number spelling (ECMAScript Number::toString)}
\usage{
format_number(x, code = "value-invalid")
}
\arguments{
\item{x}{A finite number.}

\item{code}{The refusal code for a non-finite one.}
}
\description{
Kernel section 7a number spelling (ECMAScript Number::toString)
}
