% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{jobj}
\alias{jobj}
\alias{jarr}
\title{JSON objects and arrays that stay distinct when empty}
\usage{
jobj(...)

jarr(...)
}
\arguments{
\item{...}{Members (named for an object) or elements.}
}
\description{
JSON objects and arrays that stay distinct when empty
}
