% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm15.R
\name{lm15_plain}
\alias{lm15_plain}
\title{lm15 values as plain lmcc JSON}
\usage{
lm15_plain(x)
}
\arguments{
\item{x}{An lm15 dict (from `lm15::as_dict()`) or any JSON value.}
}
\description{
Unclasses lm15's JSON objects, arrays and number tokens recursively.
}
