% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{shape_string}
\alias{shape_string}
\alias{shape_integer}
\alias{shape_number}
\alias{shape_boolean}
\alias{shape_enum}
\alias{shape_nullable}
\alias{shape_list}
\alias{shape_object}
\alias{shape_media}
\title{JSON-Schema shape builders}
\usage{
shape_string(...)

shape_integer(...)

shape_number(...)

shape_boolean(...)

shape_enum(...)

shape_nullable(shape)

shape_list(items = NULL)

shape_object(properties = NULL)

shape_media(type)
}
\arguments{
\item{...}{Extra JSON-Schema keywords (named), or enum members.}

\item{shape, items}{An inner shape.}

\item{properties}{A named list of shapes.}

\item{type}{An lm15 part type.}
}
\description{
Plain shapes for [lmcc_signature()]. Lists and objects are structured:
they need a format (section 5).
}
