% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turn.R
\name{signature_fingerprint}
\alias{signature_fingerprint}
\title{A signature's fingerprint (section 3a)}
\usage{
signature_fingerprint(sig)
}
\arguments{
\item{sig}{A signature.}
}
\description{
A signature's fingerprint (section 3a)
}
