% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_lmcc.R
\name{svd_lmcc}
\alias{svd_lmcc}
\title{Singular Value Decomposition of X and Y}
\usage{
svd_lmcc(Y_obs, sites, X, tol_eig = 0)
}
\arguments{
\item{Y_obs}{an incomplete matrix, data frame or data table to be imputed. Missing values are coded as NAs.}

\item{sites}{a numeric vector of the site locations with a length equal to the number of columns of Y_obs.}

\item{X}{a matrix, data frame or data table of covariates with the same number of rows as Y_obs.}

\item{tol_eig}{the threshold for the eigen values of the SVD decomposition.}
}
\value{
a list of objects needed for the LMCC model estimation and prediction:
\item{Y_obs}{incomplete data matrix}
\item{input}{observed methylation sites (scaled)}
\item{output}{observed data matrix}
\item{scale_sites}{methylation sites (scaled)}
\item{A}{matrices A and X stacked by columns (see reference below)}
\item{A_output_t}{matrices B and V stacked by rows (see reference below)}
\item{X}{covariates matrix}
}
\description{
Perform a SVD of the covariates matrix X and the incomplete matrix Y
}
\examples{

library(FastGP)
N = 100
K = 10
D = 2
sites = sort(runif(N))
beta = c(runif(D,2,5),runif(K-D,10,1000))
nugget = c(rep(0,D),runif(K-D,0.001,0.05))
A = matrix(runif(K*(K-D),0,0.1),nrow = K, ncol = K-D )
X = matrix(runif(D*K),ncol = D, nrow= K)
Hx = matrix(solve(t(X)\%*\%X,t(X)),nrow = length(X)/K)
A = A-X\%*\%Hx\%*\%A
A = cbind(X,A)
V = matrix(NA, nrow = K, ncol = N)
R00 = abs(outer(sites, sites, '-'))
for (d in 1:K) {
 R = matern_5_2_kernel(R00, beta = beta[d])
 R_tilde = R + nugget[d] * diag(N)
 V[d, ] = rcpp_rmvnorm_stable(1, R, rep(0, N))
}
Y_obs = A \%*\% V
obj_lmcc = svd_lmcc(Y_obs,sites,X,tol_eig = 1e-6)
}
\references{
Melina Ribaud, Aurélie Labbe and Karim Oualkacha.
Imputation in genetic methylation studies: LMCC, a Linear Model of Coregionalization with informative Covariates.
2023. hal-00000000
}
\author{
Melina Ribaud
}
