% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter.R
\name{system_msg}
\alias{system_msg}
\alias{developer_msg}
\alias{user_msg}
\alias{assistant_msg}
\alias{turns_slot}
\title{Template messages and the turn slot directive}
\usage{
system_msg(text)

developer_msg(text)

user_msg(text)

assistant_msg(text)

turns_slot(slot = "turns")
}
\arguments{
\item{text}{Template text.}

\item{slot}{A turn slot name.}
}
\description{
Template messages and the turn slot directive
}
