% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter.R
\name{use_vocab}
\alias{use_vocab}
\title{A reference to a named format or transport}
\usage{
use_vocab(name, ...)
}
\arguments{
\item{name}{The registered name.}

\item{...}{Options.}
}
\description{
A reference to a named format or transport
}
